backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
python-dotenv 